Move CoverArtThumbnail to its own file
